#55 [REFACTOR] 커스텀 훅 분리
